IRS library version for impersonate-rs
